requirements.txt
# Quantbuild E1 v1 — XAUUSD Full-Stack Trading Bot
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
pyarrow>=12.0
python-dotenv>=1.0
pydantic>=2.0

# News layer
feedparser>=6.0
httpx>=0.27

# Data sources
dukascopy-python>=4.0
yfinance>=0.2

# Optional: LLM sentiment
# openai>=1.0

# Optional: Oanda broker (live/paper trading)
# oandapyV20>=0.7

# Optional: Telegram alerts
# python-telegram-bot>=20

# Optional: Dashboard
# streamlit>=1.30

# Optional: Scheduling (live loop)
# apscheduler>=3.10

# Dev / testing
pytest>=7.0
pytest-cov
